% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{accumulation_phase}
\alias{accumulation_phase}
\title{Create accumulation phase}
\usage{
accumulation_phase(from_age, to_age, contribution = NULL, strategy = NULL)
}
\arguments{
\item{from_age}{Starting age for phase}

\item{to_age}{Ending age for phase}

\item{contribution}{ContributionFlow object}

\item{strategy}{Strategy object}
}
\value{
A Phase object
}
\description{
Create accumulation phase
}
