% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{add_market_model}
\alias{add_market_model}
\title{Add global market model}
\usage{
add_market_model(sim, model, asset = "stocks")
}
\arguments{
\item{sim}{A Simulation object}

\item{model}{A MarketModel object}

\item{asset}{Asset name}
}
\value{
Modified Simulation object
}
\description{
Add global market model
}
