% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{add_phase}
\alias{add_phase}
\title{Add a phase to the simulation}
\usage{
add_phase(sim, phase)
}
\arguments{
\item{sim}{A Simulation object}

\item{phase}{A Phase object}
}
\value{
Modified Simulation object
}
\description{
Add a phase to the simulation
}
