% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{age_to_period}
\alias{age_to_period}
\title{Helper function for age to period conversion}
\usage{
age_to_period(age, start_age, periods_per_year)
}
\arguments{
\item{age}{Age to convert}

\item{start_age}{Starting age of simulation}

\item{periods_per_year}{Number of periods per year}
}
\value{
Period number
}
\description{
Helper function for age to period conversion
}
