% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{annualize_return}
\alias{annualize_return}
\title{Annualize a periodic return}
\usage{
annualize_return(periodic_return, periods_per_year)
}
\arguments{
\item{periodic_return}{Return for a single period}

\item{periods_per_year}{Number of periods per year}
}
\value{
Annualized return
}
\description{
Annualize a periodic return
}
