% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{distribution_phase}
\alias{distribution_phase}
\title{Create distribution phase}
\usage{
distribution_phase(from_age, to_age, withdrawal = NULL, strategy = NULL)
}
\arguments{
\item{from_age}{Starting age for phase}

\item{to_age}{Ending age for phase}

\item{withdrawal}{WithdrawalFlow object}

\item{strategy}{Strategy object}
}
\value{
A Phase object
}
\description{
Create distribution phase
}
