% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cash-flows.R
\name{generate_cash_flows}
\alias{generate_cash_flows}
\title{Generic for generating cash flow amounts}
\usage{
generate_cash_flows(flow, ...)
}
\arguments{
\item{flow}{A CashFlow object}

\item{n_periods}{Number of periods}

\item{periods_per_year}{Number of periods per year}
}
\value{
Numeric vector of cash flow amounts
}
\description{
Generate cash flows for contributions (from a ContributionFlow object) or
for withdrawals (from a WithdrawalFlow object).
}
