% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market-models.R
\name{generate_returns}
\alias{generate_returns}
\title{Generic for generating returns}
\usage{
generate_returns(model, ...)
}
\arguments{
\item{model}{A MarketModel object}

\item{n_sims}{Number of simulations}

\item{n_periods}{Number of periods}

\item{periods_per_year}{Number of periods per year}
}
\value{
Matrix of returns (n_sims x n_periods)
}
\description{
Currently defined for GBMModel and HistoricalModel.
}
