% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{periodize_return}
\alias{periodize_return}
\title{Convert annual return to periodic return}
\usage{
periodize_return(annual_return, periods_per_year)
}
\arguments{
\item{annual_return}{Annual return}

\item{periods_per_year}{Number of periods per year}
}
\value{
Periodic return
}
\description{
Convert annual return to periodic return
}
