% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{periods_to_years}
\alias{periods_to_years}
\title{Convert periods to years}
\usage{
periods_to_years(periods, periods_per_year)
}
\arguments{
\item{periods}{Number of periods}

\item{periods_per_year}{Number of periods per year}
}
\value{
Number of years
}
\description{
Convert periods to years
}
