% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retirementsim-package.R
\docType{package}
\name{retirementsim-package}
\alias{retirementsim}
\alias{retirementsim-package}
\title{retirementsim: Monte Carlo Retirement Simulation Tools}
\description{
A composable, extensible R package for Monte Carlo retirement simulations using S7 classes and matrix-based computations. Provides tools for modeling retirement portfolios with various market models, cash flows, and investment strategies.
}
\author{
\strong{Maintainer}: Michael McLaren \email{m.mclaren42@gmail.com} (\href{https://orcid.org/0000-0003-1575-473X}{ORCID})

}
\keyword{internal}
