% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{sim_percentiles}
\alias{sim_percentiles}
\title{Get percentiles of simulation trajectories}
\usage{
sim_percentiles(results, probs = c(0.1, 0.25, 0.5, 0.75, 0.9))
}
\arguments{
\item{results}{A SimResults object}

\item{probs}{Probability values for percentiles}
}
\value{
Matrix of percentiles (rows = percentiles, columns = periods)
}
\description{
Get percentiles of simulation trajectories
}
