% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{sim_success_rate}
\alias{sim_success_rate}
\title{Calculate success rate of simulation}
\usage{
sim_success_rate(results, threshold = 0)
}
\arguments{
\item{results}{A SimResults object}

\item{threshold}{Minimum portfolio value to be considered successful}
}
\value{
Numeric success rate between 0 and 1
}
\description{
Calculate success rate of simulation
}
