% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{sim_summary}
\alias{sim_summary}
\title{Create summary statistics of simulation}
\usage{
sim_summary(results)
}
\arguments{
\item{results}{A SimResults object}
}
\value{
List containing summary statistics
}
\description{
Create summary statistics of simulation
}
