% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{years_to_periods}
\alias{years_to_periods}
\title{Convert years to periods}
\usage{
years_to_periods(years, periods_per_year)
}
\arguments{
\item{years}{Number of years}

\item{periods_per_year}{Number of periods per year}
}
\value{
Number of periods
}
\description{
Convert years to periods
}
