% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{holding_phase}
\alias{holding_phase}
\title{Create holding phase}
\usage{
holding_phase(from_age, to_age, strategy = NULL)
}
\arguments{
\item{from_age}{Starting age for phase}

\item{to_age}{Ending age for phase}

\item{strategy}{Strategy object}
}
\value{
A Phase object
}
\description{
Create holding phase
}
